requirements.txt
Flask
Flask-Cors
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
python-dotenv
